requirements.txt
gitpython
requests
py-cpuinfo